requirements.txt
google_api_python_client==2.179.0
google_auth_oauthlib==1.2.2
protobuf==6.32.1
python-dotenv==1.1.1
python-telegram-bot==22.4
Telethon==1.40.0
APScheduler==3.11.0
